Move test views into separate module.
